+++ components/org.wso2.carbon.identity.authenticator.saml2.sso/pom.xml
@@ -73,6 +73,10 @@
             <groupId>org.wso2.carbon.identity.carbon.auth.saml2</groupId>
             <artifactId>org.wso2.carbon.identity.authenticator.saml2.sso.common</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.identity.saml.common</groupId>
+            <artifactId>org.wso2.carbon.identity.saml.common.util</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
@@ -115,8 +119,8 @@
                             org.wso2.carbon.user.core.*; version="${carbon.kernel.package.import.version.range}",
                             org.wso2.carbon.utils.*; version="${carbon.kernel.package.import.version.range}",
                             org.wso2.carbon.identity.base; version="${identity.framework.package.import.version.range}",
-                            org.wso2.carbon.identity.core.util; version="${identity.framework.package.import.version.range}",
                             org.wso2.carbon.identity.authenticator.saml2.sso.common; version="${identity.carbon.auth.saml2.package.import.version.range}",
+                            org.wso2.carbon.identity.saml.common.util.*; version="${saml.common.util.version.range}",
                             org.joda.time; version="${joda.wso2.osgi.version.range}"
                         </Import-Package>
                         <Export-Package>
